# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "http server api docs",
        "title": "lol api docs",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/paymentHistory": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "submit information to create paymentHistory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "paymentHistory"
                ],
                "summary": "create paymentHistory",
                "parameters": [
                    {
                        "description": "paymentHistory information",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreatePaymentHistoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.CreatePaymentHistoryReply"
                        }
                    }
                }
            }
        },
        "/api/v1/paymentHistory/list": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "list of paymentHistorys by paging and conditions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "paymentHistory"
                ],
                "summary": "list of paymentHistorys by query parameters",
                "parameters": [
                    {
                        "description": "query parameters",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Params"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ListPaymentHistorysReply"
                        }
                    }
                }
            }
        },
        "/api/v1/paymentHistory/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get paymentHistory detail by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "paymentHistory"
                ],
                "summary": "get paymentHistory detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GetPaymentHistoryByIDReply"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "update paymentHistory information by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "paymentHistory"
                ],
                "summary": "update paymentHistory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "paymentHistory information",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdatePaymentHistoryByIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.UpdatePaymentHistoryByIDReply"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "delete paymentHistory by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "paymentHistory"
                ],
                "summary": "delete paymentHistory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.DeletePaymentHistoryByIDReply"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "types.Column": {
            "type": "object",
            "properties": {
                "exp": {
                    "description": "expressions, which default to = when the value is null, have =, !=, \u003e, \u003e=, \u003c, \u003c=, like",
                    "type": "string"
                },
                "logic": {
                    "description": "logical type, default value is \"and\", support \u0026, and, ||, or",
                    "type": "string"
                },
                "name": {
                    "description": "column name",
                    "type": "string"
                },
                "value": {
                    "description": "column value"
                }
            }
        },
...
        "types.CreatePaymentHistoryReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data",
                    "type": "object",
                    "properties": {
                        "id": {
                            "description": "id",
                            "type": "integer"
                        }
                    }
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
        "types.CreatePaymentHistoryRequest": {
            "type": "object",
            "properties": {
                "createAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "outTradeNo": {
                    "description": "支付订单号",
                    "type": "string"
                },
                "status": {
                    "description": "状态",
                    "type": "string"
                },
                "userPhone": {
                    "description": "用户手机号码",
                    "type": "string"
                }
            }
        },
...
        "types.DeletePaymentHistoryByIDReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data"
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
...
        "types.GetPaymentHistoryByIDReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data",
                    "type": "object",
                    "properties": {
                        "paymentHistory": {
                            "$ref": "#/definitions/types.PaymentHistoryObjDetail"
                        }
                    }
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
...
        "types.ListPaymentHistorysReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data",
                    "type": "object",
                    "properties": {
                        "paymentHistorys": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.PaymentHistoryObjDetail"
                            }
                        }
                    }
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
...
        "types.Params": {
            "type": "object",
            "properties": {
                "columns": {
                    "description": "query conditions",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Column"
                    }
                },
                "limit": {
                    "description": "lines per page",
                    "type": "integer"
                },
                "page": {
                    "description": "page number, starting from page 0",
                    "type": "integer"
                },
                "sort": {
                    "description": "sorted fields, multi-column sorting separated by commas",
                    "type": "string"
                }
            }
        },
        "types.PaymentHistoryObjDetail": {
            "type": "object",
            "properties": {
                "createAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "convert to uint64 id",
                    "type": "integer"
                },
                "outTradeNo": {
                    "description": "支付订单号",
                    "type": "string"
                },
                "status": {
                    "description": "状态",
                    "type": "string"
                },
                "userPhone": {
                    "description": "序号",
                    "type": "string"
                }
            }
        },
...
        "types.UpdatePaymentHistoryByIDReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data"
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
        "types.UpdatePaymentHistoryByIDRequest": {
            "type": "object",
            "properties": {
                "createAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "uint64 id",
                    "type": "integer"
                },
                "outTradeNo": {
                    "description": "支付订单号",
                    "type": "string"
                },
                "status": {
                    "description": "状态",
                    "type": "string"
                },
                "userPhone": {
                    "description": "序号",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type Bearer your-jwt-token to Value",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}